Feat: relocate entity checking logic from repo to service

+++ src/modules/products/products.service.ts
@@ -2,6 +2,7 @@ import { Injectable } from '@nestjs/common';
 import { UpdateProductDto } from './dto/update-product.dto';
 import { ProductsRepository } from './products.repository';
 import { ProductModel } from './models/product.model';
+import { ProductNotFoundException } from '@/common/exceptions';
 
 @Injectable()
 export class ProductsService {
@@ -20,10 +21,18 @@ export class ProductsService {
   }
 
   async update(id: number, productModel: Partial<ProductModel>) {
+    const product = await this.repo.findById(id);
+    if (!product) {
+      throw new ProductNotFoundException();
+    }
     return await this.repo.update(id, productModel);
   }
 
   async remove(id: number) {
+    const product = await this.repo.findById(id);
+    if (!product) {
+      throw new ProductNotFoundException();
+    }
     return await this.repo.remove(id);
   }
 }
